1. PLAYER UI 기능 추가 2. Material의 Texture 키 저장/로드 기능 추가 3. CRenderMgr에서 Play 카메라 + UI 카메라가 같이 작동될 수 있도록 수정 4. CRenderComponent에서 Editor 모드일 때 Dynamic Material을 보여줄 수 있는 기능 추가.

+++ Project/Client/SceneOutliner.cpp
@@ -81,6 +81,13 @@ void SceneOutliner::ShowHierarchyAll(CGameObject* pGameObject, bool afterRenew)
 
 void SceneOutliner::update()
 {
+	static bool bFirst = true;
+	if (bFirst)
+	{
+		bFirst = false;
+		Reset();
+	}
+
 	m_pSelectedScene = CSceneMgr::GetInst()->GetCurScene();
 
 	if (CEventMgr::GetInst()->HasOccurObjEvent())
@@ -451,15 +458,15 @@ bool SceneOutliner::MakePrefab()
 		return false;
 	}
 
-	if (nullptr != CResMgr::GetInst()->FindRes<CPrefab>(prefabPath))
+	if (nullptr != CResMgr::GetInst()->FindRes<CPrefab>(prefabPath) && m_bOverwritePrefab)
 	{
 		CResMgr::GetInst()->ForceDeleteRes<CPrefab>(prefabPath);
 		if (std::filesystem::exists(contentPath + prefabPath))
 		{
 			std::filesystem::remove(contentPath + prefabPath);
 		}
+		m_bOverwritePrefab = false;
 	}
-	m_bOverwritePrefab = false;
 
 	Ptr<CPrefab> pPrefab = new CPrefab{};
 	pPrefab->SetProto(m_pSelectedGameObject->Clone());
